Bug/refactor login controller v2 (#60)
State design pattern / state machines
Using the State design pattern/state machine to attempt to simplify our login journey.

The premise of this is that during such journeys, there is a finite set of states.

Defining them all, and how one transitions to another in a single place should greatly help with future maintainability & readability compared to the previous implementation.

Implementation
I've chosen an enum to achieve this: LoginJourneyState. Each value defines a nextState function, which describes how to reach the next step in the journey. Endpoints will call this function when needed, invoking the state machine which will perform actions until a state is reached that requires a redirect for the user. The endpoint can then grab this redirect from the enums other function: getLoginJourneyRedirect.

I've defined a set of functions in the login controller that should greatly simplify how we do all of the above, so any future changes to the journey (such as the migration of old COLA users) will mainly take place in this enum state machine.

If none of that makes any sense, reach out to me and I can try again!